WIP dialog sidepanel
