+++ src/background/context-menu.js
@@ -96,6 +96,8 @@ async function clickContextMenu (info = {}, tab = {}) {
   }
 
   if (info.menuItemId === openDialogMenu) {
+    chrome.sidePanel.open({ tabId: tab.id });
+    
     return trigger(eventShowDialog, {}, tab, info.frameId)
   }
