+++ src/content/dialog/dialog-footer.css
@@ -37,18 +37,24 @@
 
 /* show modals
  */
+/* 
 .briskine-dialog[modal=settings] .dialog-settings,
 .briskine-dialog[modal=actions] .dialog-actions {
+*/
+.dialog-container[modal=settings] .dialog-settings,
+.dialog-container[modal=actions] .dialog-actions {
   visibility: visible;
   opacity: 1;
   translate: 0;
   scale: 1;
 }
 
-.briskine-dialog[modal=settings] .btn-settings {
+/* .briskine-dialog[modal=settings] .btn-settings { */
+.dialog-container[modal=settings] .btn-settings {
   background: var(--border-color);
 }
 
-.briskine-dialog[modal=actions] .btn-actions {
+/* .briskine-dialog[modal=actions] .btn-actions { */
+.dialog-container[modal=actions] .btn-actions {
   background-color: var(--primary-color);
 }
