+++ src/content/dialog/dialog.js
@@ -1,36 +1,22 @@
 /* global REGISTER_DISABLED */
-import {Show, onMount, onCleanup, createSignal, createEffect, mergeProps} from 'solid-js'
 import {render} from 'solid-js/web'
+import {onMount, createSignal, createEffect} from 'solid-js'
+
+import DialogUI from './dialog-ui.js'
 
 import { eventShowDialog } from '../../config.js'
 import {
-  getTemplates,
-  getTags,
-  getAccount,
-  getExtensionData,
   on as storeOn,
   off as storeOff,
-  searchTemplates,
-  openPopup,
 } from '../../store/store-content.js'
+import autocomplete from '../autocomplete.js'
 import {isContentEditable} from '../editors/editor-contenteditable.js'
 import { isTextfieldEditor } from '../editors/editor-textfield.js'
+import { getActiveElement } from '../utils/active-element.js'
 import {bubbleTagName} from '../bubble/bubble.js'
 import {getEditableCaret, getContentEditableCaret, getDialogPosition} from './dialog-position.js'
-import autocomplete from '../autocomplete.js'
 import { getSelectionRange, setSelectionRange }  from '../utils/selection.js'
-import { getActiveElement } from '../utils/active-element.js'
 import {keybind, keyunbind} from '../keybind.js'
-import IconSearch from 'bootstrap-icons/icons/search.svg'
-import IconBriskine from '../../icons/briskine-logo-small.svg'
-
-import DialogFooter from './dialog-footer.js'
-import DialogList from './dialog-list.js'
-import DialogTemplates from './dialog-templates.js'
-import DialogSettings from './dialog-settings.js'
-import DialogActions from './dialog-actions.js'
-
-import styles from './dialog.css'
 
 function scopeElementName (name = '') {
   return `${name.toLowerCase()}-${Date.now().toString(36)}`
@@ -38,75 +24,26 @@ function scopeElementName (name = '') {
 
 let dialogInstance = null
 
-export const dialogTagName = scopeElementName('b-dialog')
-
 const modalAttribute = 'modal'
 const openAnimationClass = 'b-dialog-open-animation'
 const listSelector = '.dialog-list'
+
+export const dialogTagName = scopeElementName('b-dialog')
+
 const dialogSelector = '.briskine-dialog'
 
 function Dialog (originalProps) {
-  const props = mergeProps({
-    keyboardShortcut: '',
-    visible: false,
-  }, originalProps)
-
-  // eslint-disable-next-line no-unassigned-vars
-  let element
-
-  const [visible, setVisible] = createSignal(false)
-  const [loggedIn, setLoggedIn] = createSignal()
-  const [loading, setLoading] = createSignal()
-  const [templates, setTemplates] = createSignal([])
-  const [tags, setTags] = createSignal([])
-  const [extensionData, setExtensionData] = createSignal({})
-
-  const [searchResults, setSearchResults] = createSignal([])
-  const [searchQuery, setSearchQuery] = createSignal('')
 
   let globalAbortController = new AbortController()
   let globalListenerOptions = {
     capture: true,
     signal: globalAbortController.signal,
-  }
+  }  
 
-  let editor
-  let searchField
+  // eslint-disable-next-line no-unassigned-vars
+  let element
 
-  let cachedRange
-
-  createEffect((prev) => {
-    if (
-      visible() === true
-      && prev === false
-    ) {
-      // activate the first item in the list
-      const $list = element.querySelector(listSelector)
-      if ($list) {
-        $list.dispatchEvent(new Event('b-dialog-select-first'))
-      }
-
-      element.classList.add(openAnimationClass)
-    } else if (
-      visible() === false
-      && prev == true
-    ) {
-      element.classList.remove(openAnimationClass)
-
-      window.requestAnimationFrame(() => {
-        // clear the search query
-        if (searchField) {
-          searchField.value = ''
-        }
-
-        setSearchQuery('')
-        // close modals
-        element.removeAttribute(modalAttribute)
-      })
-    }
-
-    return visible()
-  }, false)
+  const [visible, setVisible] = createSignal(false)
 
   function show (node) {
     // dialog is already visible
@@ -190,36 +127,38 @@ function Dialog (originalProps) {
     if (removeCaretParent) {
       removeCaretParent()
     }
+  }
 
-    // give it a second before focusing.
-    // in production, the search field is not focused on some websites (eg. google sheets, salesforce).
-    setTimeout(() => {
-      searchField.focus({ preventScroll: true })
-    })
-
-    if (loading() === true) {
-      loadData()
+  function hideOnFocusout (e) {
+    if (e.target === dialogInstance) {
+      setVisible(false)
     }
   }
 
-  async function insertTemplate (id = '') {
-    await restoreSelection()
+  async function restoreSelection () {
+    // will also hide dialog because of focusout
+    editor.focus({ preventScroll: true })
 
-    // get template from cache
-    const template = templates().find((t) => t.id === id)
-
-    autocomplete({
-      template: template,
-    })
+    if (
+      isContentEditable(editor)
+      && cachedRange
+      // if the nodes change in the editor, the range containers move to the top.
+      // (e.g., linkedin message editor in shadow dom likes to re-create the
+      // entire editor dom structure on focus)
+      // this will also be true when the editor is empty, when focus() is enough.
+      && cachedRange.startContainer !== editor
+    ) {
+      await setSelectionRange(editor, cachedRange)
+    }
   }
-
+  
   function stopPropagation (e, target) {
     if (
       target
       && dialogInstance
       && (
         dialogInstance === target
-        || dialogInstance === target?.getRootNode?.()
+        || dialogInstance.shadowRoot.contains(target)
       )
     ) {
       e.stopPropagation()
@@ -234,106 +173,75 @@ function Dialog (originalProps) {
     stopPropagation(e, e.relatedTarget)
   }
 
-  async function templatesUpdated (templates = []) {
-    setTemplates(templates)
-  }
+  onMount(() => {
 
-  async function tagsUpdated (tags = []) {
-    setTags(tags)
-  }
+    element.addEventListener('b-dialog-insert', async (e) => {
+      await restoreSelection()
 
-  function extensionDataUpdated (data = {}) {
-    setExtensionData(data)
-  }
+      autocomplete({
+        template: e.detail,
+      })      
 
-  async function loadData () {
-    const extensionData = await getExtensionData()
-    extensionDataUpdated(extensionData)
-
-    setTemplates(await getTemplates())
-    setTags(await getTags())
-
-    setLoading(false)
-  }
-
-  function setAuthState () {
-    getAccount()
-    .then(() => {
-      return true
+      e.stopImmediatePropagation()
     })
-    .catch(() => {
-      return false
-    })
-    .then((status) => {
-      setLoggedIn(status)
-      setLoading(true)
+    
+    window.addEventListener('focusout', hideOnFocusout, globalListenerOptions)
+    window.addEventListener('keydown', hideOnEsc, globalListenerOptions)    
 
-      // only start loading data if the dialog is visible
-      if (visible()) {
-        loadData()
+    // prevent Gmail from handling keydown.
+    // any keys assigned to Gmail keyboard shortcuts are prevented
+    // from being inserted in the search field.
+    window.addEventListener('keydown', stopTargetPropagation, globalListenerOptions)
+    // prevent Front from handling keyboard shortcuts
+    // when we're typing in the search field.
+    window.addEventListener('keypress', stopTargetPropagation, globalListenerOptions)
+
+    // prevent parent page from handling focus events.
+    // fix interaction with our dialog in some modals (LinkedIn).
+    // prevent the page from handling the focusout event when switching focus to our dialog.
+    window.addEventListener('focusout', stopRelatedTargetPropagation, globalListenerOptions)
+    window.addEventListener('focusin', stopTargetPropagation, globalListenerOptions)    
+
+    
+    // expose show on element
+    element.show = show
+  })
+
+  createEffect((prev) => {
+    if (
+      visible() === true
+      && prev === false
+    ) {
+      // activate the first item in the list
+      const $list = element.querySelector(listSelector)
+      if ($list) {
+        $list.dispatchEvent(new Event('b-dialog-select-first'))
       }
-    })
-  }
 
-  function handleSearchFieldShortcuts (e) {
-    // only handle events from the search field
-    const target = e.composedPath()[0]
-    const $list = element.querySelector(listSelector)
-    if (
-      target !== searchField ||
-      !['Enter', 'ArrowDown', 'ArrowUp'].includes(e.key) ||
-      !$list
+      element.classList.add(openAnimationClass)
+    } else if (
+      visible() === false
+      && prev == true
     ) {
-      return
+      element.classList.remove(openAnimationClass)
+
+      window.requestAnimationFrame(() => {
+        /*
+        // clear the search query
+        if (searchField) {
+          searchField.value = ''
+        }
+
+        setSearchQuery('') 
+        */
+
+        // close modals
+        element.removeAttribute(modalAttribute)
+      })
     }
 
-    if (e.key === 'Enter') {
-      $list.dispatchEvent(new Event('b-dialog-select-active'))
-      return e.preventDefault()
-    }
-
-    let move
-    if (e.key === 'ArrowDown') {
-      move = 'next'
-    } else if (e.key === 'ArrowUp') {
-      move = 'previous'
-    }
-
-    if (move) {
-      $list.dispatchEvent(new CustomEvent('b-dialog-select', {
-        detail: move,
-      }))
-      // prevent moving the cursor to the start/end of the search field
-      e.preventDefault()
-    }
-  }
-
-  async function restoreSelection () {
-    // will also hide dialog because of focusout
-    editor.focus({ preventScroll: true })
-
-    if (
-      isContentEditable(editor)
-      && cachedRange
-      // if all the nodes are destroyed in the editor, the range resets and moves to the top.
-      // (e.g., linkedin message editor in shadow dom likes to re-create the
-      // entire editor dom structure on focus)
-      // this will also be true when the editor is empty, when focus() is enough.
-      && !(
-        cachedRange.startContainer === editor
-        && cachedRange.collapsed === true
-        && cachedRange.startOffset === 0
-      )
-    ) {
-      await setSelectionRange(editor, cachedRange)
-    }
-  }
-
-  function hideOnFocusout (e) {
-    if (e.target === dialogInstance) {
-      setVisible(false)
-    }
-  }
+    return visible()
+  }, false)
 
   function hideOnEsc (e) {
     if (e.key === 'Escape' && visible()) {
@@ -345,133 +253,7 @@ function Dialog (originalProps) {
 
       restoreSelection()
     }
-  }
-
-  onMount(() => {
-    // check authentication state
-    setAuthState()
-    storeOn('login', setAuthState)
-    storeOn('logout', setAuthState)
-
-    storeOn('templates-updated', templatesUpdated)
-    storeOn('tags-updated', tagsUpdated)
-    storeOn('extension-data-updated', extensionDataUpdated)
-
-    let searchDebouncer
-    searchField = element.querySelector('input[type=search]')
-
-    // search for templates
-    searchField.addEventListener('input', (e) => {
-      if (searchDebouncer) {
-        clearTimeout(searchDebouncer)
-      }
-
-      const searchValue = e.target.value
-      if (searchValue) {
-        searchDebouncer = setTimeout(async () => {
-          const {query, results} = await searchTemplates(searchValue)
-          if (query === searchValue) {
-            setSearchQuery(searchValue)
-            setSearchResults(results)
-          }
-        }, 50)
-      } else {
-        setSearchQuery('')
-      }
-    })
-
-    // keyboard navigation and insert for templates
-    window.addEventListener('keydown', handleSearchFieldShortcuts, globalListenerOptions)
-    element.addEventListener('b-dialog-insert', (e) => {
-      insertTemplate(e.detail)
-      e.stopImmediatePropagation()
-    })
-
-    element.addEventListener('click', (e) => {
-      const target = e.target
-
-      // open and close modals
-      const btnModalAttribute = 'data-b-modal'
-      const modalBtn = target.closest(`[${btnModalAttribute}]`)
-      if (modalBtn) {
-        const modal = modalBtn.getAttribute(btnModalAttribute)
-        if (element.getAttribute(modalAttribute) !== modal) {
-          element.setAttribute(modalAttribute, modal)
-        } else {
-          element.removeAttribute(modalAttribute)
-
-          // focus the search field when closing the modals,
-          // and returning to the list view.
-          if (searchField) {
-            searchField.focus({ preventScroll: true })
-          }
-        }
-
-        return
-      }
-
-      // login button
-      if (target.closest('.dialog-login-btn')) {
-        e.preventDefault()
-        openPopup()
-      }
-    })
-
-    window.addEventListener('focusout', hideOnFocusout, globalListenerOptions)
-    window.addEventListener('keydown', hideOnEsc, globalListenerOptions)
-
-    // prevent parent page from handling composed events.
-    // fix interaction with our dialog in some modals (Gmail, LinkedIn new post, GitHub search).
-    const stopTargetEvents = [
-      // stop Gmail from preventing keys assigned to Gmail shortcuts to be typed
-      // inside the dialog search field
-      'keydown',
-      'keyup',
-      // stop Front from catching keyboard shortcuts when typing in the dialog
-      'keypress',
-      // stop GitHub search from closing the dialog when clicking inside it
-      'mousedown',
-      'mouseup',
-      // stop GitHub search from preventing opening the dialog
-      // stop LinkedIn New Post from stealing focus when opening the dialog
-      'focus',
-      'focusin',
-    ]
-
-    const stopRelatedTargetEvents = [
-      // stop GitHub search and LinkedIn new post from handling blur
-      'blur',
-      'focusout',
-    ]
-
-    stopTargetEvents.forEach((event) => {
-      window.addEventListener(event, stopTargetPropagation, globalListenerOptions)
-    })
-
-    stopRelatedTargetEvents.forEach((event) => {
-      window.addEventListener(event, stopRelatedTargetPropagation, globalListenerOptions)
-    })
-
-    // expose show on element
-    element.show = show
-  })
-
-  onCleanup(() => {
-    storeOff('login', setAuthState)
-    storeOff('logout', setAuthState)
-
-    storeOff('templates-updated', templatesUpdated)
-    storeOff('tags-updated', tagsUpdated)
-    storeOff('extension-data-updated', extensionDataUpdated)
-
-    globalAbortController.abort()
-
-    globalAbortController = new AbortController()
-    globalListenerOptions = {
-      capture: true,
-      signal: globalAbortController.signal,
-    }
-  })
+  }  
 
   return (
     <div
@@ -481,63 +263,12 @@ function Dialog (originalProps) {
         'briskine-dialog-visible': visible(),
       }}
       >
-      <style>{styles}</style>
-      <div
-        classList={{
-          'dialog-container': true,
-          'dialog-safari': REGISTER_DISABLED,
-        }}
-        tabindex="-1"
-        >
-
-        <div class="dialog-search">
-          <input type="search" value="" placeholder="Search templates..." spellcheck="false" />
-          <div class="dialog-search-icon">
-            <IconSearch />
-          </div>
-        </div>
-
-        <div class="dialog-content">
-          <Show when={!loggedIn()}>
-            <div class="dialog-info d-flex">
-              <div class="dialog-info-icon">
-                <IconBriskine />
-              </div>
-              <div>
-                <a href="" class="dialog-login-btn">Sign in</a> to Briskine to access your templates.
-              </div>
-            </div>
-          </Show>
-
-          <Show
-            when={searchQuery()}
-            fallback={(
-              <DialogTemplates
-                loggedIn={loggedIn()}
-                loading={loading()}
-                templates={templates()}
-                tags={tags()}
-                extensionData={extensionData()}
-                />
-            )}
-            >
-            <DialogList
-              loggedIn={loggedIn()}
-              list={searchResults()}
-              showTags={extensionData().dialogTags}
-              tags={tags()}
-              />
-          </Show>
-        </div>
-
-        <Show when={loggedIn()}>
-          <DialogFooter shortcut={props.keyboardShortcut} />
-          <DialogSettings extensionData={extensionData()} />
-          <DialogActions />
-        </Show>
-      </div>
-    </div>
-  )
+        <DialogUI 
+          keyboardShortcut={this.keyboardShortcut} 
+          visible={visible()} 
+        />
+    </div> 
+  ) 
 }
 
 customElements.define(dialogTagName, class extends HTMLElement {
