+++ src/manifest.json
@@ -18,7 +18,8 @@
     "contextMenus",
     "storage",
     "scripting",
-    "unlimitedStorage"
+    "unlimitedStorage",
+    "sidePanel"
   ],
   "host_permissions": [
     "https://*/*",
@@ -52,7 +53,8 @@
       "icons/icon-32.png",
       "icons/icon-32-loggedout.png",
       "icons/icon-48.png",
-      "icons/icon-48-loggedout.png"
+      "icons/icon-48-loggedout.png",
+      "sidepanel/main.html"
     ],
     "matches": [
       "https://*/*",
@@ -70,6 +72,9 @@
       "32": "icons/icon-32.png"
     },
     "default_popup": "popup/popup.html"
+  },
+  "side_panel": {
+    "default_path": "sidepanel/main.html"
   },
   "content_security_policy": {
     "extension_pages": "script-src 'self'; object-src 'self'; frame-ancestors 'none'",
