+++ src/sidepanel/main.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>Briskine</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <style type="text/css" >
+      .briskine-dialog {
+        /* height: 100% !important; */
+        display: block !important;
+      }
+      #app, #app-body {
+        height: 100%;
+      }
+    </style>
+  </head>
+  <body>
+    <div id="app" class="briskine-dialog"></div>
+    <script src="sidepanel.js" type="module"></script>    
+  </body>
+</html>
