+++ webpack.config.js
@@ -94,6 +94,7 @@ function extensionConfig ({ mode, safari, manifest, firebaseConfig}) {
     new CopyWebpackPlugin({
       patterns: [
         { from: 'src/popup/popup.html', to: 'popup/' },
+        { from: 'src/sidepanel/main.html', to: 'sidepanel/' },
         { from: 'src/icons/', to: 'icons/' },
         { from: 'src/content/sandbox/sandbox.html', to: 'sandbox/' },
         { from: 'LICENSE', to: '' }
@@ -129,6 +130,7 @@ function extensionConfig ({ mode, safari, manifest, firebaseConfig}) {
     entry: {
       background: './src/background/background.js',
       popup: './src/popup/popup.js',
+      sidepanel: './src/sidepanel/sidepanel.js',
       content: {
         import: './src/content/index.js',
         // force iife
